+++ src/core/graph.rs
@@ -2,8 +2,9 @@ pub use super::edge::Edge;
 pub use super::node::Node;
 use super::edgeinformation::EdgeInformation;
 use super::routing::RoutingResult;
-use super::routingoptions::AlgorithmOptions;
+use super::routing::options::RoutingAlgorithmOptions;
 pub use super::weight::WeightCalculator;
+
 use std::collections::HashSet;
 
 use std::rc::Rc;
@@ -26,8 +27,13 @@ pub trait Graph {
     where
         F: FnMut(i32, &Rc<Edge>); //TODO this needs to become independent of actual edge storage
 
-    fn route(&self, opts: &AlgorithmOptions<Self>, start: i32, end: i32) -> Option<RoutingResult>
+    fn route(&self, opts: &RoutingAlgorithmOptions<Self>, start: i32, end: i32) -> Option<RoutingResult>
     where
+        Self:Sized;
+
+    //a vector of sets of nodeids, each set is a strongly connected subgraph
+    fn get_strongly_connected_subgraphs(&self, opts: &RoutingAlgorithmOptions<Self>) -> Vec<HashSet<i32>> 
+    where 
         Self:Sized;
 
     //functions with default implementations